Add support for stable rust
Resolves https://github.com/vcombey/fallible_collections/issues/1

Unstable features now require the `unstable` feature to be enabled.
Most functionality is preserved, with the exception of the btree module
and the TryClone trait.

A dependency on the hashbrown crate is added to use its CollectionAllocErr.
It will be used more extensively when support for types requiring fallible
allocation is added.

See https://github.com/vcombey/fallible_collections/issues/2.